# backend/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Isi Stasiun Backend API
  version: 1.0.0
  description: REST API for station attributes, analytics, survey ingestion, pipeline callbacks, transparency, authentication, premium analysis, and AI copilot routing.
servers:
  - url: /
    description: Current backend origin
...
paths:
...
  /api/v1/stations:
    get:
      tags: [Stations]
      summary: List stations
      operationId: listStations
      parameters:
        - $ref: "#/components/parameters/AreaTypeQuery"
        - name: operator
          in: query
          schema: {type: string}
          description: Exact operator name filter
      responses:
        "200":
          $ref: "#/components/responses/StationList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stations/{id}:
    get:
      tags: [Stations]
      summary: Get station detail
      operationId: getStation
      parameters:
        - $ref: "#/components/parameters/StationIDPath"
      responses:
        "200":
          description: Station detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data: {$ref: "#/components/schemas/Station"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /api/v1/stations/{id}/entrances:
    get:
      tags: [Stations]
      summary: List station entrances
      operationId: listStationEntrances
      parameters:
        - $ref: "#/components/parameters/StationIDPath"
      responses:
        "200":
          description: Entrance list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: {$ref: "#/components/schemas/Entrance"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /api/v1/analytics/spending-gap:
    get:
      tags: [Analytics]
      summary: List spending gap estimates
      operationId: listSpendingGap
      responses:
        "200": {$ref: "#/components/responses/SpendingGapList"}
        "500": {$ref: "#/components/responses/InternalError"}
  /api/v1/analytics/spending-gap/{station_id}:
    get:
      tags: [Analytics]
      summary: List spending gap estimates for a station
      operationId: listStationSpendingGap
      parameters:
        - $ref: "#/components/parameters/StationIDNamedPath"
      responses:
        "200": {$ref: "#/components/responses/SpendingGapList"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "500": {$ref: "#/components/responses/InternalError"}
  /api/v1/analytics/category-gap:
    get:
      tags: [Analytics]
      summary: List unmet business categories
      operationId: listCategoryGap
      parameters:
        - $ref: "#/components/parameters/StationIDQuery"
      responses:
        "200":
          description: Category gap list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: {$ref: "#/components/schemas/CategoryGap"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "500": {$ref: "#/components/responses/InternalError"}
  /api/v1/analytics/rent-flow-index:
    get:
      tags: [Analytics]
      summary: List rent-flow index results
      operationId: listRentFlowIndex
      parameters:
        - $ref: "#/components/parameters/StationIDQuery"
      responses:
        "200":
          description: Rent-flow index list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: {$ref: "#/components/schemas/RentFlowIndex"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "500": {$ref: "#/components/responses/InternalError"}
  /api/v1/analytics/event-potential:
    get:
      tags: [Analytics]
      summary: List event activation potential
      operationId: listEventPotential
      parameters:
        - $ref: "#/components/parameters/StationIDQuery"
      responses:
        "200":
          description: Event potential list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: {$ref: "#/components/schemas/EventPotential"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /api/v1/copilot/query:
    post:
      tags: [Copilot]
      summary: Forward a natural-language query to the AI service
      operationId: queryCopilot
      description: Public but rate-limited. Upstream failures currently return a stable fallback answer with HTTP 200.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CopilotQueryRequest"}
      responses:
        "200":
          description: AI response or degraded fallback
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data: {$ref: "#/components/schemas/CopilotQueryResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-Service-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AreaTypeQuery:
      name: area_type
      in: query
      schema: {$ref: "#/components/schemas/AreaType"}
    StationIDPath:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    StationIDNamedPath:
      name: station_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    StationIDQuery:
      name: station_id
      in: query
      schema: {type: string, format: uuid}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    StationList:
      description: Station list
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/SuccessEnvelope"
              - type: object
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Station"}
    SpendingGapList:
      description: Spending gap rows
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/SuccessEnvelope"
              - type: object
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/SpendingGap"}
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: {type: boolean, enum: [true]}
        message: {type: string}
        data: {}
    ErrorEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: {type: boolean, enum: [false]}
        message: {type: string}
        data: {nullable: true}
...
    AreaType:
      type: string
      enum: [residential, office, mixed]
    TimeSlot:
      type: string
      enum: [morning, midday, evening, night]
    Category:
      type: string
      enum: [makanan_minuman, ritel_kemasan, apotek_kesehatan, jasa, lainnya]
    Station:
      type: object
      required: [id, name, code, operator, area_type, latitude, longitude, entrance_count]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        code: {type: string}
        operator: {type: string}
        area_type: {$ref: "#/components/schemas/AreaType"}
        latitude: {type: number, format: double, minimum: -90, maximum: 90}
        longitude: {type: number, format: double, minimum: -180, maximum: 180}
        entrance_count: {type: integer, minimum: 0}
    Entrance:
      type: object
      required: [id, label, latitude, longitude]
      properties:
        id: {type: string, format: uuid}
        label: {type: string}
        latitude: {type: number, format: double}
        longitude: {type: number, format: double}
    SpendingGap:
      type: object
      required: [station_id, station_name, potential_low_p10, potential_high_p90, captured_low_p10, captured_high_p90, gap_low_p10, gap_high_p90, time_slot, computed_at]
      properties:
        station_id: {type: string, format: uuid}
        station_name: {type: string}
        potential_low_p10: {type: number, format: double}
        potential_high_p90: {type: number, format: double}
        captured_low_p10: {type: number, format: double}
        captured_high_p90: {type: number, format: double}
        gap_low_p10: {type: number, format: double}
        gap_high_p90: {type: number, format: double}
        time_slot: {$ref: "#/components/schemas/TimeSlot"}
        computed_at: {type: string, format: date-time}
    CategoryGap:
      type: object
      required: [station_id, category, demand_in_area, available_in_station]
      properties:
        station_id: {type: string, format: uuid}
        category: {$ref: "#/components/schemas/Category"}
        demand_in_area: {type: boolean}
        available_in_station: {type: boolean}
    RentFlowIndex:
      type: object
      required: [plot_id, station_id, offered_rent, measured_flow, index, is_outlier]
      properties:
        plot_id: {type: string}
        station_id: {type: string, format: uuid}
        offered_rent: {type: number, format: double}
        measured_flow: {type: number, format: double}
        index: {type: number, format: double}
        is_outlier: {type: boolean}
    EventPotential:
      type: object
      required: [station_id, zone_id, activation_score, recommended_slot]
      properties:
        station_id: {type: string, format: uuid}
        zone_id: {type: string}
        activation_score: {type: number, format: double}
        recommended_slot: {type: string, nullable: true}
...
    CopilotQueryRequest:
      type: object
      required: [query]
      properties:
        query: {type: string, minLength: 1, maxLength: 500, example: "Stasiun mana yang memiliki spending gap tertinggi?"}
        station_id: {type: string, format: uuid}
    CopilotQueryResponse:
      type: object
      required: [answer]
      properties:
        answer: {type: string}
        suggested_layers:
          type: array
          items: {type: string}
        spatial_filter:
          type: object
          additionalProperties: true
...